+++ packages/fl_select/lib/src/select/cascading_select.dart
@@ -748,8 +748,6 @@ class CascadingSelectState extends State<CascadingSelect> {
                 SelectActionBar(
                   resetText: delegate.resetText,
                   applyText: delegate.applyText,
-                  resetFlex: delegate.actionBarTheme?.resetFlex,
-                  applyFlex: delegate.actionBarTheme?.applyFlex,
                   onResetTap: _onResetTap,
                   onApplyTap: _onApplyTap,
                 ),
@@ -888,8 +886,6 @@ class CascadingSelectState extends State<CascadingSelect> {
               SelectActionBar(
                 resetText: delegate.resetText,
                 applyText: delegate.applyText,
-                resetFlex: delegate.actionBarTheme?.resetFlex,
-                applyFlex: delegate.actionBarTheme?.applyFlex,
                 onResetTap: _onResetTap,
                 onApplyTap: _onApplyTap,
               ),