+++ packages/fl_select/lib/src/select/expandable_select.dart
@@ -299,8 +299,6 @@ class _ExpandableSelectState extends State<ExpandableSelect> {
               SelectActionBar(
                 resetText: delegate.resetText,
                 applyText: delegate.applyText,
-                resetFlex: delegate.actionBarTheme?.resetFlex,
-                applyFlex: delegate.actionBarTheme?.applyFlex,
                 onResetTap: _onResetTap,
                 onApplyTap: _onApplyTap,
               ),