+++ packages/fl_select/lib/src/select/list_select.dart
@@ -159,8 +159,6 @@ class ListSelectState extends State<ListSelect> {
               SelectActionBar(
                 resetText: delegate.resetText,
                 applyText: delegate.applyText,
-                resetFlex: delegate.actionBarTheme?.resetFlex,
-                applyFlex: delegate.actionBarTheme?.applyFlex,
                 onResetTap: _onResetTap,
                 onApplyTap: _onApplyTap,
               ),