+++ packages/fl_select/lib/src/select/side_nav_select.dart
@@ -559,8 +559,6 @@ class SideNavSelectState extends State<SideNavSelect> {
             SelectActionBar(
               resetText: delegate.resetText,
               applyText: delegate.applyText,
-              resetFlex: delegate.actionBarTheme?.resetFlex,
-              applyFlex: delegate.actionBarTheme?.applyFlex,
               onResetTap: _onResetTap,
               onApplyTap: _onApplyTap,
             ))