+++ packages/fl_select/lib/src/select/tab_nav_select.dart
@@ -280,8 +280,6 @@ class TabNavSelectState extends State<TabNavSelect> {
                 SelectActionBar(
                   resetText: delegate.resetText,
                   applyText: delegate.applyText,
-                  resetFlex: delegate.actionBarTheme?.resetFlex,
-                  applyFlex: delegate.actionBarTheme?.applyFlex,
                   onResetTap: _onResetTap,
                   onApplyTap: _onApplyTap,
                 ),
@@ -375,8 +373,6 @@ class TabNavSelectState extends State<TabNavSelect> {
               SelectActionBar(
                 resetText: delegate.resetText,
                 applyText: delegate.applyText,
-                resetFlex: delegate.actionBarTheme?.resetFlex,
-                applyFlex: delegate.actionBarTheme?.applyFlex,
                 onResetTap: _onResetTap,
                 onApplyTap: _onApplyTap,
               ),